backend/requirements.txt
# ── Core framework ───────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0

# ── Hardware detection ───────────────────────────────────────────────────
psutil>=5.9.0

# ── Storage ─────────────────────────────────────────────────────────────
chromadb>=0.5.0           # persistent vector/document storage

# ── NLP / ML (optional — needed for semantic search / embeddings) ────────
# sentence-transformers>=3.0.0   # uncomment for semantic code search
# torch>=2.3.0                   # needed by sentence-transformers

# ── Web / browsing ───────────────────────────────────────────────────────
playwright>=1.44.0         # headless browser for computer-use
beautifulsoup4>=4.12.0
lxml>=5.2.0

# ── Sandbox / execution ──────────────────────────────────────────────────
# pyautogui>=0.9.54        # uncomment for computer-use mouse/keyboard
# xvfbwrapper>=0.2.9       # uncomment for headless display (Linux)

# ── Image / vision (optional) ────────────────────────────────────────────
Pillow>=10.3.0

# ── Git integration ──────────────────────────────────────────────────────
GitPython>=3.1.43

# ── Utilities ────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=23.2.1
tenacity>=8.3.0            # retries for network calls
PyYAML>=6.0.1
